requirements.txt
certifi==2020.12.5
olefile==0.46
pytesseract
pytest
requests
Pillow==7.1.2